bss hack
save bss start and end so i can zero them in the go runtime init code. can't do
it in the bootloader because there isn't enough room.

it would make memory management easier if the bootloader loaded the kernel
segments into physically contiguous memory and then setup a page directory that
mapped the virtual addresses to the physical addresses. then we could also do
the bss zeroing before loading.

since we are low on room though, the bootloader would need to be more than one
sector and thus we would have to change the bootloader to load the rest of
itself. i'll do this later.