finish segmentation and thus TLS setup
took me many hours to figure out why far ret to fix CS was faulting. since the
default operand size of long mode is 32-bits, lret exptected the far pointer to
be one quad on the top of the stack that looked like: 0x0000SSSSPPPPPPPP where
S are the segment bits and P are the address bits.

i was pushing two quads, first the segment then the address. that only works if
you use lret with REX.W prefix (lretq).