fix evil non-deterministic bug
because the go kernel runs in ring 0 and grows its own stack, the stack while
handling a timer interrupt may not be mapped in every process. thus, during a
context switch, the ret instruction in the lcr3 function faults. because the
stack is no longer valid, a triple-fault/reboot results.

one fix is to inline lcr3. that way the stack of the interrupted task is not
used after the new page map is loaded.