SMP
find and initialize cpus. the test boots up all found cpus (controllable with
CPUS env variable) each of which print a hello message, spin for a moment, and
then fault to make sure the IDT/GDT are properly setup.

there is no synchronization yet. next step is to make the runtime/kernel
thread-safe.