build a biscuit fs
don't assume that the fs is on a separate disk, making it easier to boot
biscuit on real hardware. therefore we need a mechanism to figure out where the
fs blocks start. instead of using a real MBR, just write the bits into a
special variable in the go kernel binary.