inode transactions via inode daemons
the basic idea is that some transaction operation is sent to one of the
involved idaemons. this idaemon operates on the inode state and then sends (and
waits for completion of) a new operation to another involved idaemon which may
itself generate and send a new operation, etc. if the transaction fails
somewhere, each idaemon cleans up and returns the error.

the main challenges with this design are avoiding these deadlocks:

- when simultaneously traversing up and down the directory tree, two idaemons
  may both block sending to each other. solve this by having one
  channel/goroutine daemon per direction (however i think i can make traversing
  up the tree unnecessary altogether, so this is not yet implemented).

- when handling a transaction operation that results in an idaemon sending
  operations to itself. instead of using a goroutine per operation per idaemon,
  deal with this by partitioning the operations into simple/transactional
  operations. both classes are handled by a separate goroutine. transactional
  operations can only generate/send simple operations while simple operations
  cannot themselves generate any operations. (however, directory tree
  traversing is a special case: deadlock cannot occur since each idaemon
  forwards operations its children instead of sending an operation to an
  arbitrary idaemon).