prevent kernel threads from polling interrupt queue until timer interrupt
timeslice size was a red-herring -- the syscall latency problem was that once
all syscall work was completed, all kernel threads just spin in trap() looking
for interrupts in the queue until the timeslice expires, only then finally
returning to user-space (only to immediately syscall again).

explicitly yield after Gosched to prevent this. this is not a great solution
because, if there are many runnable user processes, an unlucky kernel thread
may yield and start running a user process instead of handling the syscall.
maybe this doesn't matter much.

it would probably be better to cooperatively switch to trap() -- maybe we can
do this in the stack growth prologue.