don't start sysmon() goroutine -- biscuit doesn't need it
i noticed that sysmon() gets scheduled very frequently while i was benchmarking
the cost of channels. sysmon() is called by a background goroutine started by
the runtime. it does the following:
	- if enabled, dump some scheduling debug info
	- reclaim P's from syscalls entered via cgo or the Syscall package
	  (syscalls made by the runtime are not treated specially)
	- reclaim P's from goroutines that have run for more than 10ms
	- poll the network after some timeout
	- forcing a GC after some timeout
	- forcing a "scavenge" (calling madvise(DONT_NEED) on all pages not
	  recently used) of the heap after some timeout

we probably don't care about any of this but the GC, which biscuit can do
better since it knows when to do it. thus, remove sysmon from biscuit.

alternatively, i could fix biscuit's concept of time so that sysmon() sleeps
for the correct amount of time.