wakeup a goroutine for CPU interrupts/no polling
biscuit starts a special trap consuming goroutine, as before, but now instead
of the Gosched/Usleep nonsense, it calls Trapsched which puts the goroutine to
sleep if there are no more interrupts to process. when an interrupt comes in,
the handler sets a flag which is checked by all Ps when looking for goroutines
to run. they race for ownership of the trap consuming goroutine and the winner
puts it on his run queue.

another notable change: when an M can't find a goroutine to run, it yields
instead of stopping itself.

the difference is immediately noticable -- the test hardware with 1 or 8 CPUs
is much more responsive. unfortunately, it doesn't improve our numbers in the
bmwrite benchmark! interesting.

the sleep/wakeup code is a little complicated; i hope to simplify it.

comment out my addition to the compiler that generates the check in the
function prologue since we don't need it. it may be useful later.