keep files in memory after unlinked on disk
this change was more intrusive than i expected. because open fails under
various circumstances (ie O_CREAT on a directory), the checking code must
decrement the in memory file's reference count in the presence of a failure.
however, the checking code was not part of the lookup/open transaction, so such
a scheme would not be correct. thus the checking code was moved into the inode
daemon so the check/ref decrement are atomic with the lookup/open.

also prevent opening directories as writable, chdiring to files, and add
O_DIRECTORY flag.