close files, decrement memory ref count
thus open files are usable after they are unlinked from the file system and
will be freed on close.

the code does not deal with the case where the machine crashes in between an
open file being unlinked and the file's close(). i plan on logging such files
in the journal and removing the journal entry once the file has been freed
after close().