figure out LAPIC clock speed using 8254 PIT
we need a sense of time because i want timeouts to be specified in seconds,
milliseconds, etc. instead of cycles. since the 8254 has a well known frequency
we can use it to determine the LAPIC's frequency (which is different from the
CPU's frequency), which we can then use to keep track of nanoseconds per timer
interrupt.

qemu's lapic clock is 1Ghz while the test hardware's is 132Mhz.

time package is still broken -- this is the first step to fixing it.