use TSC for timekeeping instead of LAPIC clock
LAPIC clock is annoying because the current time is unknown inbetween when a
time interrupt is generated but before the time quantum is added to the clock
(since the current count register has been reset!).  also, LAPIC timers are not
synchronized between CPUs (obviously).

TSC is much easier since it increases constantly on recent hardware. Intel
guarantees that TSC increases at a constant rate from Pentium 4s and on.