fake syscall to enable profiling
make time(1) call it.

its convenient to extract the profile from biscuit to a computer with go tools
to vizualize the profile data. the method of extracting the profile is very
sophisticated: biscuit dumps the profile to the serial console in a format
suitable for "xxd -r" so we can copy and paste it.