use runtime memmove/memclr instead of homerolled ones
the builtin ones are optimized to used SSE and increase performance of a hacked
version of mailbench by 57%.

i made my own when i first started biscuit because i wasn't sure what the
built-in ones depended on and we need these very early in the boot process.

thus, we must call fpuinit() earlier in boot process.