fix process fd table races
although the fd table is protected by a lock, the old code (which was adapted
from pre-thread times) would create an fd with with the lock held, but would
update some of the fd's fields outside the lock. thus in the presence of a
concurrent fork, the forking thread may observe an fd with uninitialized
fields, resulting in a page fault.