fix races between copying from/to user virtual addresses and page faults
if two threads page fault and call sys_read concurrently, the old sys_read code
may copy the result to a page that was marked copy-on-write and thus has been
since replaced. this code serializes copies to/from user space with page faults
and adds explicit checks to fault-in the page that will be the destination of
the copy if it is marked copy-on-write (to prevent a parent/child from
corrupting the memory of the other). these explicit checks can go away once
biscuit can handle a syscall without switching to the kernel pmap.

care is taken so that a process's page map lock is only held for the duration
of the copy (to avoid holding the lock while waiting for blocks from disk).

as biscuit's mailbench performance doubles over and over, races that were
hidden by sluggishness get exposed.