keep a process's VM info in vmregion_t/vmseg_t
biscuit's previous two approaches to VM management were both biased towards
either small or large address spaces. this approach is fast for both.

track each region of mapped memory in proc_t.vmregion, which is a linked list.
inserts into a vmregion are coalesced in order to avoid long interations during
lookup. removals from a vmregion cause splits. since we now know the first
virtual address of every mapped region, we can use a slice to track virtual
pages instead of a map.

there are two main benefits to this approach: 1) when forking, we can scan/copy
only the necessary PTEs -- we never examine empty PTEs. 2) we can track
allocated pages with a slice instead of a map which requires less CPU time and
no allocations for insertion or iteration.

biscuit is now as fast as linux on the forkonly benchmark! this improvement
helps us a lot on all the other benchmarks too.