test hardware bugfix: race between ifree() and icreate()
the race occurs when icreate() is given an inode block that is shared with
another inode that is about to be freed. before icreate() is able to install
its inode into the block, ifree() decides to free the inode block since it
doesn't see icreate()'s new inode.

thank you test hardware.