mkbdisk.py: fill unused blocks with crap to help detect FS bugs
i've been baffled all morning wondering why a bug that apparently has nothing
to do with concurrency exhibits only on the hardware and not qemu. the answer
is that, because qemu determines the emulated HDD size by the size of the file,
i was zero filling all unused blocks. i do not initialize the hardware's HDD
since that would take a long time (and should be pointless). thus an FS bug
that uses an uninitialized block may do the right thing on qemu but not the
hardware.