fix FS segment bug; cleanup segment code
also fix an amusing, ancient bug from when i apparently didn't yet know that
the calling convention for both runtime C and Go code is to pass return values
on the stack. my assembly code expected the return value in rax. magically,
this code never broke because the compiler is apparently very likely to use rax
to hold the return value in order to copy it to the caller's stack frame.