bugfix: prevent deadlock with TLB shootdown+GC
the old TLB shootdown code would wait for all shootdowns to complete by cas'ing
the variable holding the number of outstanding shootdowns with 0. it was
possible that, before the CPU waiting for all shootdowns to complete could
observe the 0 count, another CPU would install a new TLB count.

i thought this was fine since the waiting CPU must observe 0 at some point.
turns out it may not if the CPU that installs the new count is preempted by a
GC before it actually sends the IPIs. then the GC loops forever, trying to
preempt the waiting CPU (which is not preemptable as it is looping for the
count to reach 0).

this new code ensures that the waiting CPU figures out that its shootdown is
complete, even if some other CPU installs a new count before the waiting CPU
observes a 0 count.