malloc: Don't forget to account for the header while allocating memory
At the moment, malloc() calls mmap() to allocate exactly as much as the user
asks (after rounding up the size to be a multiple of page-size, of course).
But after that, it embeds a 32-byte header within this allocated space, which
implies that the user doesn't always have as much usable memory as he asked
for. So fix this by taking the header size into consideration appropriately.