finish futex(2), pthread_cond_* &c
most of the complexity comes from the optimization to prevent a thundering herd
after pthread_cond_broadcast(3).

the original design was simple for lock/unlock only; it has since been thrown
out due to complexity. the original design did not spuriously wake up threads,
but avoiding spurious wake ups seems to require a lot more locked instructions
from kernel space and allowing threads to be woken up both with and without the
lock acquired depending on the circumstances. this made the timeout code
(necessary for pthread_cond_timedwait(3)) quite tricky since it may be hard to
figure out whether or not you have the lock after a timeout.