replace pipe daemon with typical implementation
the pipe daemon code is not a good opportunity for us to compare the trade-offs
of CSP style because the daemon approach has no benefits here (not even
simplicity!), but has some serious and demonstrable performance problems
(mainly due to the misuse of channels as locks). thus, the poor performance of
the pipe daemon has more to do with a bad design and less to do with CSP style.

the new code, unlike the daemon code, requires an order of magnitude less
scheduling events, less memory copying, and less heap allocations. it doubles
redis throughput.