make malloc(3)/free(3) less brain-dead
my stupidly simple malloc/free track how many objects are allocated on a
segment and munmaps the segment when the count reaches 0. if a program mallocs
and frees the same object in a loop, malloc will mmap/munmap a page for a
single object continuously. apparently redis-benchmark does exactly this.

thus, modify free s.t. the segment containing the object being freed is
unmapped only if it is not the current segment into which objects are being
allocated.

improves redis throughput by 47%. on redis, linux is only 1.77x faster than
biscuit now!