bootloader: load kernel via int 13h instead of PIO
this change makes it easier to boot biscuit on other PCs that have different
disk controllers. it also allows biscuit to use an AHCI disk driver since the
disk controller is no longer required to be in IDE mode.

the change is complicated by the fact that the biscuit kernel is > 3MB, but int
13h is a real-mode BIOS call and thus can only address 1MB of memory. thus the
bootloader must switch between protected and real modes, reading disk blocks to
a real-mode buffer and then copying to the actual destination after switching
to protected mode.