eagerly free user pages on process termination
otherwise a process which tries to allocate much memory may fail, even if there
is enough memory since a lot of memory may be sitting in the pmap of a
terminated process which some idle CPU still has loaded. freeing the pmap
itself is still delayed.

it's harder for biscuit to detect when per-CPU resources are no longer in use
since, for performance, we do not change pmaps during goroutine context
switches.