poll sleeping Ms, too
although i just modified the scheduler to poll for runnable threads, the old
code wouldn't also check for sleeping threads that should be woken. thus these
threads would not be awoken until the next timer interrupt (which is, on
average, 500us away). a runtime thread sometimes sleeps for 100us when
synchronizing with other runtime threads.

this fixes our nginx and mailbench multicore performance. nginx with 1, 2, and
4 cores have throughput within 5% of linux!