limit number of user process VM objects to prevent OOM
a funny situation can occur if a process creates the maximum number of VM
objects, but isn't able to remember both the starting address of the mapping
and the length. since unmapping the middle of an existing mapping creates a new
VM object, i cannot allow the process to partially unmap a regions if the VM
object limit has been reached.  thus a user program may not be able to easily
free memory, even if it wants too!