limit maximum number of segments remembered by a TCP socket to prevent OOM
a program which calls many write(2)s with a small amount of data can reach the
TX limit. when this occurs, we collapse all outstanding segments to the same
timestamp; thus we will transmit more than strictly necessary if the segments
aren't ACK'ed soon.

the RX limit can be reached if a connection receives many non-contiguous
segments out of order. in this case, we forget that we received the farthest
segments in the stream.