use page allocator for pipe buffers and cleanup
which are reused for UNIX domain sockets, too. the error handling is much
easier if we lazily allocate the page since then the read(2)/write(2) code
paths have to deal with it with an allocation failure; read/write already have
robust error handling since they operate on untrusted user addresses.