bugfix: prevent disk read/log commit race
with the old code, disk reads would read stale data when reading an evicted
vnode page that had been written to the FS log but before the log had been
committed.

i considered a few ways to fix this; the following is the simplest i've come up
with so far: any disk reads for vnodes that have been recently evicted need to
see if the log contains a more recent version of the block. once the log has
been committed, the vnode no longer needs to check the log.

it's tempting to simply wait for the log to commit before evicting a vnode's
pages, but this may result in deadlock since the log may not be able to commit
until the allocating FS operation completes.