Add a flag to mmapi to ask to increase the refcount on the physical page backing an mmaped file.  On a pagefault, if it is is file page, the handler calls mmapi with this flag set to 1.  If we just want the info about the file, then mmapi is called with false (the default).  This avoids the following race: if a file is immediately evicted from the buffer cache before mmapi returns (e.g., on the relse()), then the file page is returned to the page allocator before its refcount is increased and mapped.
Test the flag by evicting every file block immediately.  Without the flag, we
see the race.  With this fix, usertests passes, except for mmaptest(). This is a
bug in mmaptest.  Biscuit doesn't guarantees to write mmaped files through to
the disk blocks backing the file, so no reason that 'Z' shows up later when we
open the file.  Maybe Biscuit's mmap should be better.