memory reservation mechanism
a goroutine reserves n bytes by subtracting n from the global reservation
counter. if the global reservation counter is less than n, the reservation
fails. when the goroutine allocates, it subtracts the number of bytes allocated
from its reservation.

when the system call/interrupt is over, the goroutine returns the remaining
bytes by adding them to the global reservation counter.

after the GC sweeps all pages, the number of newly swept bytes are added back
to the global reservation counter.