bugfix: deadlock in O_CREAT
this code attempted to lock the parent directory and target file directly,
without deadlock avoidance (iref_lockall).

the old FS had to lock the parent directory and target file simultaneously in
order to guarantee that the target file was not unlinked.

with the fancy new FS, we can guarantee the target file isn't unlinked by
simply getting a reference to the target imemnode_t before unlocking the
parent. thus there is no need for iref_lockall.

discovered via usertests on bhw