make UNIX SOCK_DGRAM socket receives atomic, for restart-ability
it is not good if a restarted recvmsg(2) call by a non-hog thread mysteriously
returns a zero length message with a valid sender address.

recvmsg(2) is still broken though, because writing the sender's socket address
size and ancilary data size to user memory is not atomic w.r.t. the
data/address copy.