prepare for reliable kill
CSP can make killing even more annoying! we got a bit lucky with how we use
channels; there are only four code sites where a goroutine can block for a long
time on a channel operation, and the other three are simple.

there are two ways that CSP makes killing more annoying. first, if the hog
goroutine is blocking on a channel on which other non-hog goroutines are also
blocked, we need to somehow tell the non-hogs to retry. second, many operations
in CSP style involve two or more messages between the two parties.  well the
code is unlikely to cope with the other party mysteriously bailing before all
messages have been sent or received.

the futex code was an example of the second case. luckily it is easy to fix by
having the futex daemon send wakeups asynchronously. that way, if the hog is
sleeping on a futex and gets killed before the futex wakes it, the futex daemon
won't block forever.

in biscuit, both of these annoyances are naturally dealt when using conditional
variables instead of channels.