Avoid deadlock when several processes are updating the same file.
op_begin and inode lock must be done in the same order, but write acquires inode
lock and then asks for log admission.  Other processes calling close may have
been admitted to the log, but cannot acquire the inode lock that the writer
holds.

for write, i think, it is fine to delay acquiring inode lock until after the
write has been admitted to the log.

there is probably a similar deadlock possible with ifree(). we need to check all
system calls that don't do log admission at the top of the call.  (maybe a
simple tool, even though we have only a handful cases now.)