Remove the dead inode list from icache, which is a source of contention in addDead() and op_end_and_free().  Instead, on refdown return if inode should be freed, and at the end of rename/link/unlink/close check if an inode should be freed (in its own op).  This changes avoids the global dead list, but cannot use defer in a few places because now need to check the result of refdown.
This change exposed a bug in link/unlink/open that was introduced when switching
from inums to *imemnode_t. icreat() should return pointer to inode instead of
inum so caller has a reference to an inode so that the inode won't be deleted
(and/or recreated with same inum) while the caller is using it.

Passes usertests, ufs_test, and parrun 4. Parrun -d 10 4: 16401 msg/s