Fix races between unlink/evict and lock-free lookup.
Uping a reference count can return failure if the refcounted object has been
evicted. ilookup_lockfree fails in that case, since it is racing with an
evict. Uping a reference count also returns its previous value. If the previous
value is 0, ilookup_lockfree also fails, because it is racing with an unlink.

Namei() asks ilookup_free to up the reference count only for the final target
inode (i.e., not for intermediate inodes on the path).  Thus, namei() may be
looking at stale inodes, but in this particular case it is ok, i
believe. Namei() will ensure that it returns a valid reference for the target
inode.

Maybe I missed a race?

Performance of parrun -d 10 4: 16455 msg/s