remove per-imemnode parent lists
the parent lists are difficult to maintain efficiently and without deadlocking
with eviction (must update the parent lists of both the parent and children
dirents of an evicted directory inode).

make a dirent's pointer to an imemnode not increase the refcount of the target
imemnode. because unopened, unused files now have a refcount of 0, lock-free
namei must check that the target imemnode's link count is non-zero since the it
was looked-up without locking its parent, which may have been concurrently
unlinked.