bugfix: make writes to ICR for TLB shootdowns are atomic
an unlucky context switch between the two writes could cause the IPI to be sent
to the wrong CPU, which results in the sender waiting forever for an
acknowledgement.

this bug didn't cause any problems when broadcasting TLB shootdowns because the
value of one of the writes (the destination APIC ID) was unused.